In HDF5 1.6.4 rdcc_nelmts is "int" rather than "unsigned long" in the call to H5Pget_cache
git-svn-id: https://svn.nexusformat.org/code/trunk@466 ff5d1e40-2be0-497f-93bd-dc18237bd3c7

+++ src/napi5.c
@@ -127,7 +127,7 @@ NXstatus CALLING_STYLE NX5closegroup (NXhandle fid);
   unsigned int vers_major, vers_minor, vers_release, am1 ;
   hid_t fapl;     
   int mdc_nelmts;
-  unsigned long rdcc_nelmts;
+  int rdcc_nelmts;
   size_t rdcc_nbytes;
   double rdcc_w0;
